Add files via upload
PHP_Pet-Shop
One of the projects I worked on was a Pet Shop Management System. The goal of this project was to create an efficient and user-friendly system to manage various aspects of a pet shop. The system I developed included features for managing inventory, customer information, and sales transactions. It also allowed the shop to keep track of pet details, such as breed, age, and health records. I used PHP for the backend development, ensuring the system was robust and scalable. For the frontend, I utilized HTML and CSS to create an intuitive interface for users. The system also incorporated a database to securely store all the necessary information.